Fix issue with some properties for named styles
Named styles can be defined on the worbook level or on the worksheet level. The ExcelStyle constructor took this into account, but some of the properties addressed the wrong collection.

This is fixed by making the correct `ExcelXfs` source object a private property for the `ExcelStyle` object; it remains available after the constructor has finished.